changed render clamp implementation to be integer

+++ sola-engine/platform/browser/src/main/java/technology/sola/engine/platform/browser/core/BrowserCanvasRenderer.java
@@ -61,7 +61,7 @@ public class BrowserCanvasRenderer implements Renderer {
   }
 
   @Override
-  public void setClamp(float x, float y, float width, float height) {
+  public void setClamp(int x, int y, int width, int height) {
     throw new NotYetImplementedException();
   }
